% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_island_tbl_missing_species.R
\name{add_island_tbl_missing_species}
\alias{add_island_tbl_missing_species}
\title{Calculates the number of missing species to be assigned to each island clade
in the island_tbl object and assigns the missing species to them. In the case
that multiple genera are in an island clade and each have missing species
the number of missing species is summed. Currently the missing species
are assigned to the genus that first matches with the missing species table,
however a more biologically or stochastic assignment is in development.}
\usage{
add_island_tbl_missing_species(missing_species, missing_genus, island_tbl)
}
\arguments{
\item{missing_species}{A two column data frame of genera with missing species
and the number of missing species}

\item{missing_genus}{A list of character vectors containing the genera in
each island clade}

\item{island_tbl}{Object of Island_tbl class from DAISIEprep package}
}
\value{
Object of Island_tbl class
}
\description{
Calculates the number of missing species to be assigned to each island clade
in the island_tbl object and assigns the missing species to them. In the case
that multiple genera are in an island clade and each have missing species
the number of missing species is summed. Currently the missing species
are assigned to the genus that first matches with the missing species table,
however a more biologically or stochastic assignment is in development.
}
\examples{
\dontrun{
  #WIP
}
}
