% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_species.R
\name{add_missing_species}
\alias{add_missing_species}
\title{Adds a specified number of missing species to an existing island_tbl at the
colonist specified by the species_name argument given. The species given is
located within the island_tbl data and missing species are assigned.}
\usage{
add_missing_species(island_tbl, num_missing_species, species_name)
}
\arguments{
\item{island_tbl}{Object of Island_tbl class from DAISIEprep package}

\item{num_missing_species}{Numeric specifying how many missing species to add
to an island colonist}

\item{species_name}{Character string determining which island colonist the
missing species are assigned. This is done by matching the species name
given to the species within each island colonist, therefore the species name
given does not need to match the clade_name of the island colonist (which is
a representative species for the clade)}
}
\value{
Object of Island_tbl class (see DAISIEprep package)
}
\description{
Adds a specified number of missing species to an existing island_tbl at the
colonist specified by the species_name argument given. The species given is
located within the island_tbl data and missing species are assigned.
}
\examples{
\dontrun{"WIP"}
}
