% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bird_ddl_complete_ds_asr}
\alias{bird_ddl_complete_ds_asr}
\title{DAISIE data list of birds of Madagascar (complete data).}
\format{
A list with 3 elements:
\describe{
  \item{colonist_name}{character string with a species representative of the
  island clade}
  \item{branching_times}{numeric vector with the island age, time of island
  colonisation, and any branching times if the colonist has speciated on the
  island}
  \item{stac}{numeric determining the endemicity status and colonisation
  time precision of the island colonist}
  \item{type1or2}{numeric determining whether the island colonist is a type 1
  or type 2 clade, type 1 is the default state, type 2 is when those clades
  have a distinct set of rates}
}
}
\usage{
bird_ddl_complete_ds_asr
}
\description{
A dataset containing the colonist name, colonisation and branching time,
missing species, and whether the island colonist is a type 1 or 2.
}
\keyword{datasets}
