% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist_stats.R
\name{checklist_stats}
\alias{checklist_stats}
\title{Calculates summary statistics for the checklists}
\usage{
checklist_stats(checklist)
}
\arguments{
\item{checklist}{An island species checklist}
}
\value{
A list of summary stats
}
\description{
Calculates summary statistics for the checklists
}
