% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_daisie_output.R
\name{collate_daisie_output}
\alias{collate_daisie_output}
\title{Function reads in each of the DAISIE maximum likelihood outputs stored in the
specified results folder and extracts the best maximum likelihood optimisation
for each different phylogeny and stores it as tabular data}
\usage{
collate_daisie_output(results_dir)
}
\arguments{
\item{results_dir}{A character string of the folder name in
inst/daisie_results to read the results from}

\item{oceanic_or_nonoceanic}{A character string, either "oceanic" or
"nonoceanic"}

\item{num_phylos}{Numeric of how many phylogenies were used}
}
\value{
tibble
}
\description{
Function reads in each of the DAISIE maximum likelihood outputs stored in the
specified results folder and extracts the best maximum likelihood optimisation
for each different phylogeny and stores it as tabular data
}
\examples{
# name of results of results folder
results_dir <- "amp_ddl_complete_ds_asr"
# number of phylogeny replicates used
num_phylos <- 3
# oceanic results
}
