% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_daisie_transition_matrix.R
\name{create_daisie_transition_matrix}
\alias{create_daisie_transition_matrix}
\title{Create the transition matrix for the DAISIE data extraction using the
ASR algorithm}
\usage{
create_daisie_transition_matrix(
  zero_endemic_ext = TRUE,
  zero_endemic_back_col = TRUE
)
}
\arguments{
\item{zero_endemic_ext}{Set the extinction rate of endemic species to zero.
Default is `TRUE`, if `FALSE` then the endemic species (state 3) can go to
state 1 (not present on the island).}

\item{zero_endemic_back_col}{Set the back colonisations rate of endemic
species to zero. Default is `TRUE`, if `FALSE` then the endemic species
(state 3) can go to state 2 (non-endemic on the island).}
}
\value{
A 3x3 integer matrix.
}
\description{
Creates a 3x3 matrix with the constraints on the ancestral state
reconstruction when fitting the Mk model ([castor::asr_mk_model()]) using
[DAISIEprep::add_asr_node_states()].

The constraints are colonisation to the island is from state 1
(`not_present`) to state 2 (`nonendemic`) and species cannot colonise as
endemics. Island extinction is constrained to `nonendemic` to `not_present`
(state 2 to state 1), island extinction of `endemic` cannot happen (state 3
to state 1), back-colonisation from the island is not allowed (state 3 to
state 2), and anagenesis has its own rate (state 2 to state 3).
}
\examples{
create_daisie_transition_matrix
}
