% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation to be inherited by other functions in the package}
\usage{
default_params_doc(bar)
}
\arguments{
\item{bar}{temporary argument name}
}
\value{
Nothing
}
\description{
Documentation to be inherited by other functions in the package
}
\keyword{internal}
