% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_species.R
\name{extract_species}
\alias{extract_species}
\title{Wrapper function for reading in data, processing raw dat, counting number of
missing species, formatting data, extracting island data, inputting missing
species that have phylogenetic data and outputting the island data, the
missing species that have not been assigned and the phylogenetic data used
to extract the species.}
\usage{
extract_species(
  checklist_file_name,
  phylo_file_name,
  dna_or_complete,
  daisie_status,
  extraction_method,
  rate_model
)
}
\arguments{
\item{checklist_file_name}{Character string with the name of the checklist
file (and its file extension) to load}

\item{phylo_file_name}{Character string with the name of the phylogenetic
data (and its file extension) to load}

\item{dna_or_complete}{A character string, either "DNA" or "complete" (case
insensitive) to determine whether to produce data for the DNA-only phylogeny
or the complete phylogeny}

\item{daisie_status}{Boolean determining whether to use the true endemicity
status or the augmented endemicity status, called the daisie status, which
adjusts for species that are a product of the island processes but have since
migrated}

\item{extraction_method}{Character string, either "min" or "asr" determining
whether to use the `min` or `asr` algorithms for the
`DAISIEprep::extract_island_species()` extraction process, see `DAISIEprep`
documentation for more information.}

\item{rate_model}{A character string determining the discrete character
Markov model (Mk) to be used. Only used when `extraction_method = "asr"`.
Options are equal-rates `"ER"`, all rates different `"ARD"`, symmetrical
`"SYM"`, only stepwise transitions `"SUEDE"` and only stepwise transitions
to neighbouring states `"SRD".`}
}
\value{
List with:
  * `multi_island_tbl`, the island community data
  * `phylods`, the phylogenetic data with associated endemicity data
  * `no_island_tbl_missing_species`, the species that have not been assigned to
  the island community data
}
\description{
Wrapper function for reading in data, processing raw dat, counting number of
missing species, formatting data, extracting island data, inputting missing
species that have phylogenetic data and outputting the island data, the
missing species that have not been assigned and the phylogenetic data used
to extract the species.
}
