% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{madagascar_birds_dna}
\alias{madagascar_birds_dna}
\title{Name and endemicity status of the birds of Madagascar (DNA-only data).}
\format{
A data frame with 163 rows and 2 variables:
\describe{
  \item{tip_labels}{character string with underscore separation of genus and
  species name}
  \item{tip_endemicity_status}{character string of endemicity status, all
  lowercase}
  ...
}
}
\usage{
madagascar_birds_dna
}
\description{
A dataset containing the binomial taxonomic names (genus and species name)
of the amphibian species that are present on the island of Madagascar.
}
\keyword{datasets}
