% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{madagascar_mammals}
\alias{madagascar_mammals}
\title{Name and endemicity status of the mammals of Madagascar.}
\format{
A data frame with 215 rows and 2 variables:
\describe{
  \item{tip_labels}{character string with underscore separation of genus and
  species name}
  \item{tip_endemicity_status}{character string of endemicity status, all
  lowercase}
  ...
}
}
\usage{
madagascar_mammals
}
\description{
A dataset containing the binomial taxonomic names (genus and species name)
of the mammal species that are present on the island of Madagascar.
}
\keyword{datasets}
