% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{madagascar_squamates_complete}
\alias{madagascar_squamates_complete}
\title{Name and endemicity status of the squamata of Madagascar (complete data).}
\format{
A data frame with 389 rows and 2 variables:
\describe{
  \item{tip_labels}{character string with underscore separation of genus and
  species name}
  \item{tip_endemicity_status}{character string of endemicity status, all
  lowercase}
  ...
}
}
\usage{
madagascar_squamates_complete
}
\description{
A dataset containing the binomial taxonomic names (genus and species name)
of the squamate species that are present on the island of Madagascar.
}
\keyword{datasets}
