% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_median_rd_params.R
\name{min_median_rd_params}
\alias{min_median_rd_params}
\title{Takes the collated output from the DAISIE model and calculates which
parameter estimates that are closest to the median (i.e. smallest relative
deviation from the median for all parameter)}
\usage{
min_median_rd_params(data)
}
\arguments{
\item{data}{Collated daisie output}
}
\value{
tibble
}
\description{
Takes the collated output from the DAISIE model and calculates which
parameter estimates that are closest to the median (i.e. smallest relative
deviation from the median for all parameter)
}
