% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_col_daisie_data_list.R
\name{plot_col_daisie_data_list}
\alias{plot_col_daisie_data_list}
\title{Plots the distribution of colonisation times (points) across the history
of an island for each taxonomic group given in the list of daisie_data_list
objects}
\usage{
plot_col_daisie_data_list(listed_daisie_data_list, col_time = "stem")
}
\arguments{
\item{listed_daisie_data_list}{Named list of daisie_data_list objects}

\item{col_time}{A character string, either "stem" or "crown"}
}
\value{
ggplot2 object
}
\description{
Plots the distribution of colonisation times (points) across the history
of an island for each taxonomic group given in the list of daisie_data_list
objects
}
\examples{
data("amp_ddl_complete_ds_asr")
data("vm_ddl_complete_ds_asr")
data("squa_ddl_complete_ds_asr")
daisie_data_list <- list(
  Amphibians = amp_ddl_complete_ds_asr,
  Mammals = vm_ddl_complete_ds_asr,
  Squamates = squa_ddl_complete_ds_asr
)
plot_col_daisie_data_list(listed_daisie_data_list = daisie_data_list)
}
