% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_col_island_tbl.R
\name{plot_col_island_tbl}
\alias{plot_col_island_tbl}
\title{Plots the distribution of colonisation times (points) across the history
of an island for each taxonomic group given in the list of multi_island_tbl
objects}
\usage{
plot_col_island_tbl(multi_island_tbl_list, col_time = "stem")
}
\arguments{
\item{multi_island_tbl_list}{Named list of daisie_data_list objects}

\item{col_time}{A character string, either "stem" or "crown"}
}
\value{
ggplot2 object
}
\description{
Plots the distribution of colonisation times (points) across the history
of an island for each taxonomic group given in the list of multi_island_tbl
objects
}
\examples{
amphibians <- readRDS(file = system.file(
  "extdata",
  "extracted_data",
  "amp_data",
  "amp_it_complete_ds_asr.rds",
  package = "MadIsland",
  mustWork = TRUE
))

mammals <- readRDS(file = system.file(
  "extdata",
  "extracted_data",
  "vm_data",
  "vm_it_complete_ds_asr.rds",
  package = "MadIsland",
  mustWork = TRUE
))

squamates <- readRDS(file = system.file(
  "extdata",
  "extracted_data",
  "squa_data",
  "squa_it_complete_ds_asr.rds",
  package = "MadIsland",
  mustWork = TRUE
))

multi_island_tbl_list <- list(
  Amphibians = amphibians,
  Mammals = mammals,
  Squamates = squamates
)
plot_col_island_tbl(multi_island_tbl_list = multi_island_tbl_list)
}
