% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{process_raw_data}
\alias{process_raw_data}
\title{Processes raw data file (checklist of island species) which can contain
data on many aspects (e.g. phenotypic trait) down to a data file with only
the species names in the phylogeny and their endemicity status on the island}
\usage{
process_raw_data(file_name, dna_or_complete, daisie_status)
}
\arguments{
\item{file_name}{character string with the file name, including the file
extension (needs to be .csv)}

\item{dna_or_complete}{A character string, either "DNA" or "complete" (case
insensitive) to determine whether to produce data for the DNA-only phylogeny
or the complete phylogeny}

\item{daisie_status}{Boolean determining whether to use the true endemicity
status or the augmented endemicity status, called the daisie status, which
adjusts for species that are a product of the island processes but have since
migrated}
}
\value{
Data frame
}
\description{
Processes raw data file (checklist of island species) which can contain
data on many aspects (e.g. phenotypic trait) down to a data file with only
the species names in the phylogeny and their endemicity status on the island
}
\examples{
data <- process_raw_data(
  file_name = "volant_mammal_checklist.csv",
  dna_or_complete = "complete",
  daisie_status = TRUE
)
}
