% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_checklist.R
\name{read_checklist}
\alias{read_checklist}
\title{Read a checklist into R given a file name}
\usage{
read_checklist(file_name)
}
\arguments{
\item{file_name}{character string with the file name, including the file
extension (needs to be .csv)}
}
\value{
Data frame
}
\description{
Read a checklist into R given a file name
}
\examples{
read_checklist(file_name = "vm_checklist.csv")
}
