% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_island_tbl_missing_species.R
\name{rm_island_tbl_missing_species}
\alias{rm_island_tbl_missing_species}
\title{Loops through the genera that have missing species and removes the ones that
are found in the missing genus list which have phylogenetic data. This is
useful when wanting to know which missing species have not been assigned to
the island_tbl using `DAISIEprep::add_multi_missing_species()`.}
\usage{
rm_island_tbl_missing_species(missing_species, missing_genus)
}
\arguments{
\item{missing_species}{A two column data frame of genera with missing species
and the number of missing species}

\item{missing_genus}{A list of character vectors containing the genera in
each island clade}
}
\value{
Data frame
}
\description{
Loops through the genera that have missing species and removes the ones that
are found in the missing genus list which have phylogenetic data. This is
useful when wanting to know which missing species have not been assigned to
the island_tbl using `DAISIEprep::add_multi_missing_species()`.
}
\examples{
\dontrun{
  #WIP
}
}
