% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_daisie_fit.R
\name{summarise_daisie_fit}
\alias{summarise_daisie_fit}
\title{Takes the collated output from the DAISIE model and calculates the mean and
standard deviation of the log likelihood and Bayesian information criterion}
\usage{
summarise_daisie_fit(data)
}
\arguments{
\item{data}{Collated daisie output}
}
\value{
tibble
}
\description{
Takes the collated output from the DAISIE model and calculates the mean and
standard deviation of the log likelihood and Bayesian information criterion
}
