% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_daisie_rates.R
\name{summarise_daisie_rates}
\alias{summarise_daisie_rates}
\title{Takes the collated output from the DAISIE model and calculates the mean and
standard deviation of each of the five or six model parameters}
\usage{
summarise_daisie_rates(data)
}
\arguments{
\item{data}{Collated daisie output}
}
\value{
tibble
}
\description{
Takes the collated output from the DAISIE model and calculates the mean and
standard deviation of each of the five or six model parameters
}
