% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.R
\name{summarise_data}
\alias{summarise_data}
\title{Calculates the summary statistics for island community data that has been
extracted and formatted as an `Island_tbl` class.}
\usage{
summarise_data(data)
}
\arguments{
\item{data}{An `Island_tbl` or list of `Island_tbl` objects}
}
\value{
List of numerics
}
\description{
Calculates the summary statistics for island community data that has been
extracted and formatted as an `Island_tbl` class.
}
\examples{
amphibians <- readRDS(file = system.file(
  "extdata",
  "extracted_data",
  "amp_data",
  "amp_it_complete_ds_asr.rds",
  package = "MadIsland",
  mustWork = TRUE
))
summarise_data(data = amphibians)
}
