% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_missing_species.R
\name{unique_missing_species}
\alias{unique_missing_species}
\title{Determines the unique genera that are included in the island clades contained
within the island_tbl object and stores them as a list with each genus
only occuring once in the first island clade it appears in}
\usage{
unique_missing_species(island_tbl)
}
\arguments{
\item{island_tbl}{Object of Island_tbl class from DAISIEprep package}
}
\value{
list of character vectors
}
\description{
Determines the unique genera that are included in the island clades contained
within the island_tbl object and stores them as a list with each genus
only occuring once in the first island clade it appears in
}
\examples{
\dontrun{
  #WIP
}
}
