% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_missing_species.R
\name{count_missing_species}
\alias{count_missing_species}
\title{Reads in the checklist of all species on an island, including those that are
not in the phylogeny (represented by NA) and counts the number of species
missing from the phylogeny each genus}
\usage{
count_missing_species(file_name)
}
\arguments{
\item{file_name}{character string with the file name, including the file
extension (needs to be .csv)}
}
\value{
Data frame
}
\description{
Reads in the checklist of all species on an island, including those that are
not in the phylogeny (represented by NA) and counts the number of species
missing from the phylogeny each genus
}
\examples{
count_missing_species(file_name = "mammal_checklist.csv")
}
