% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_daisie_transition_matrix.R
\name{create_daisie_transition_matrix}
\alias{create_daisie_transition_matrix}
\title{Create the transition matrix for the DAISIE data extraction using the
ASR algorithm}
\usage{
create_daisie_transition_matrix(...)
}
\arguments{
\item{...}{[dots] Not used, warns if arguments are supplied.}
}
\value{
A 3x3 integer matrix.
}
\description{
Creates a 3x3 matrix with the constraints on the ancestral state
reconstruction when fitting the Mk model ([castor::asr_mk_model()]) using
[DAISIEprep::add_asr_node_states()].

The constraints are colonisation to the island is from state 1
(`not_present`) to state 2 (`nonendemic`) and species cannot colonise as
endemics. Island extinction (state 3 and state 2 to state 1) are constrained
to the same rate, back-colonisation from the island (state 3 to state 2)
is given its own rate, as is anagenesis (state 2 to state 3).
}
\examples{
create_daisie_transition_matrix
}
