% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Documentation to be inherited by other functions in the package}
\usage{
default_params_doc(
  bar,
  file_name,
  checklist,
  island_tbl,
  missing_species,
  missing_genus,
  dna_or_complete,
  checklist_file_name,
  phylo_file_name,
  daisie_status,
  extraction_method,
  num_missing_species,
  species_name
)
}
\arguments{
\item{bar}{temporary argument name}

\item{file_name}{character string with the file name, including the file
extension (needs to be .csv)}

\item{checklist}{data frame with information on species on the island}

\item{island_tbl}{Object of Island_tbl class from DAISIEprep package}

\item{missing_species}{A two column data frame of genera with missing species
and the number of missing species}

\item{missing_genus}{A list of character vectors containing the genera in
each island clade}

\item{dna_or_complete}{A character string, either "DNA" or "complete" (case
insensitive) to determine whether to produce data for the DNA-only phylogeny
or the complete phylogeny}

\item{checklist_file_name}{Character string with the name of the checklist
file (and its file extension) to load}

\item{phylo_file_name}{Character string with the name of the phylogenetic
data (and its file extension) to load}

\item{daisie_status}{Boolean determining whether to use the true endemicity
status or the augmented endemicity status, called the daisie status, which
adjusts for species that are a product of the island processes but have since
migrated}

\item{extraction_method}{Character string, either "min" or "asr" determining
whether to use the `min` or `asr` algorithms for the
`DAISIEprep::extract_island_species()` extraction process, see `DAISIEprep`
documentation for more information.}

\item{num_missing_species}{Numeric specifying how many missing species to add
to an island colonist}

\item{species_name}{Character string determining which island colonist the
missing species are assigned. This is done by matching the species name
given to the species within each island colonist, therefore the species name
given does not need to match the clade_name of the island colonist (which is
a representative species for the clade)}
}
\value{
Nothing
}
\description{
Documentation to be inherited by other functions in the package
}
\keyword{internal}
