% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MadIsland-package.R
\docType{package}
\name{MadIsland-package}
\alias{MadIsland}
\alias{MadIsland-package}
\title{MadIsland: Package to hold all the scripts to reproduce paper on colonisation and divesification history of Madagascar's tetrapods}
\description{
Reproduce all analyses and plots from the paper on the colonisation and diversification history of Madagascar's tetrapods. Note that this package is tailored to use the Hábrók High Performance Computing Cluster (HPCC) \url{https://wiki.hpc.rug.nl/habrok/start} therefore rerunning certain analyses sequentially on a desktop/laptop will be time consuming.
}
\keyword{internal}
