% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_raw_data.R
\name{process_raw_data}
\alias{process_raw_data}
\title{Processes raw data file (checklist of island species) which can contain
data on many aspects (e.g. phenotypic trait) down to a data file with only
the species names in the phylogeny and their endemicity status on the island}
\usage{
process_raw_data(file_name)
}
\arguments{
\item{file_name}{character string with the file name, including the file
extension (needs to be .csv)}
}
\value{
Data frame
}
\description{
Processes raw data file (checklist of island species) which can contain
data on many aspects (e.g. phenotypic trait) down to a data file with only
the species names in the phylogeny and their endemicity status on the island
}
\examples{
data <- process_raw_data(file_name = "mammal_checklist.csv")
}
